Update test cases for API request
